+++ src/main/core/support/configuration.rs
@@ -453,13 +453,6 @@ pub struct ExperimentalOptions {
     #[clap(help = EXP_HELP.get("unblocked_vdso_latency").unwrap().as_str())]
     pub unblocked_vdso_latency: Option<units::Time<units::TimePrefix>>,
 
-    /// Enable extended YAML conventions (merge keys and extension fields). Can only be enabled on
-    /// the command line (enabling in the configuration file is a no-op).
-    #[clap(hide_short_help = true)]
-    #[clap(long, value_name = "bool")]
-    #[clap(help = EXP_HELP.get("use_extended_yaml").unwrap().as_str())]
-    pub use_extended_yaml: Option<bool>,
-
     /// The host scheduler implementation, which decides how to assign hosts to threads and threads
     /// to CPU cores
     #[clap(hide_short_help = true)]
@@ -514,7 +507,6 @@ impl Default for ExperimentalOptions {
                 units::TimePrefixUpper::Sec,
             ))),
             strace_logging_mode: Some(StraceLoggingMode::Off),
-            use_extended_yaml: Some(false),
             scheduler: Some(Scheduler::ThreadPerCore),
         }
     }